Initial diff-api Puppet Server plugin
This commit adds the skeleton of a module for Puppet Server that will
eventually plug in functionality for catalog diffing. The plugin provides
a new Trapperkeeper service that mounts an API endpoint at
`puppet/v3/diff-catalog`. Calls to this endpoint borrow a JRuby interpreter
and then dispatch to a Ruby handler under `lib/puppet_x`.

Most of this code was recycled from the Sharpie/puppet-facts_upload module.