Add puppet module skeleton
This commit adds the skeleton of a puppet module. Applying the the
catalog_diff_api::server class to a node will configure Puppet Server
to mount the plugin JAR during startup and sets auth.conf rules for
allowing access to the API endpoint.

A Rakefile is also present that provides tasks for building the JAR, packaging
the module and running Beaker testsuites.