(CDPE-542) Allow configuration for fresh compile masters
Previous to this commit, the module could not be added to the PE Master
node group, as a new compile master being spun up would not have the
pe_server_version fact (the fact is laid down by the pe release
package).

This commit changes it to treat the lack of a fact as a fresh install of
a compile master and continue on, with a hard require on the
Package['pe-puppetserver'] resource. This should prevent it from being
applied to a non compile master.