(CDPE-556) Add custom node classifier indirector
Previous to this commit, the catalog compilation was using the default
node terminus of :classifier. This impl would lookup the trusted facts
from the local cert, which we are not presenting due to impersonating
another node.

This commit copies that indirector here and modifies just the fact
lookup to trust what comes from PDB.