(CDPE-592) Use node.add_server_facts to properly set server_facts
Previous to this commit, the code for injecting server side facts into
the node was using node.merge, however that function does not end up
setting the @server_facts variable of the node, and just merges the
facts into the base fact set. This would lead customers using the
$server_facts variable to get compilation errors.
This commit fixes it by using the nodes add_server_facts method that
will take care of properly setting the variables for us.