(CDPE-1195) Convert the secret_key to an empty string
Previous to this commit, we would pass nil to the save_storage_settings
endpoint. However in CD4PE 2.6.0, the endpoint started encrypting the
secret regardless of provider. If the customer is using DISK as their
provider, then they will not provide a secret, causing a server error.
This commit converts the secret_key from nil to an empty string as a
temp work around until 2.7.0 is released