(CDPE-1199) Fix port mapping of cd4pe container
Previous to this commit, the port mapping of the cd4pe container was
remapping the internal port as well as the external, instead of leaving
the exposed internal ports alone. This meant that if a customer
attempted to define a non-default port, then the services would no
longer be available.
This commit fixes it by leaving the internal ports as what they should
be when changing the external ports