(CDPE-1566) Fix deps between cd4pe and root_config
Previously, when cd4pe and cd4pe::root_config were applied
simultaneously, a deadlock could occur wherein CD4PE could not be
installed because it could not be configured, because it was not
running.

This commit creates a more intelligent dependency ordering allowing
CD4PE to be first installed, then configured, then refreshed if
necessary.