CDPE-1583 adds logic to install shared job hw in root_config
This change adds a couple functions to find out whether shared/root job hw exists on the current cd4pe installation and to create new agent credentials if there aren't any for the root user. Then, we hand off to the pipelines module to install the agent if needed. This is to make the time to use even shorter by installing hw for the user