CDPE-1583 adds wait and wraps create_new_agent_credentials in begin/rescue
Currently, the pipelines::agent class is being called too quickly after the docker container is refreshed, so the agent download is getting a 500 response. this change adds a 5 second wait (in the form of an exec) before the module tries to install the agent. this also adds a begin/rescue to the function that creates new agent credentials, so compilation doesn't fail if it fails