CDPE-1583 fixes bug with ssl settings
Currently, we try to save SSL settings in the root configuration regardless of whether the user has put anay in. The backend, however, requires values for all the fields so sending an empty payload to the server causes an error to be returned. This adds a conditional in the provider that does not try to save settings if no values have been passed to the resource