CDPE-1583 fixes unit tests
in order to comply with the line length convention, I was separating
strings (mock session cookies) over multiple lines using the + operator. This apparently adds a
CR/LF to a string and you cannot have CR/LFs in HTTP headers, so this
was causing the methods under test to fail and ultimately the tests. To
avoid this, you can use '\' to separate strings over lines instead.