CDPE-1583 adds conditional in root_config/reduce wait
Currently, if the CD4PE server cannot be reached for whatever reason, or the  is undefined, compilation fails spectacularly. Thus, if  is undef, we should just skip trying to install the agent. This change adds that conditional in the root_config class. Also, this change shortens the 'wait' in the root_config class to 1. The shorter wait time still serves its purpose and lowers impact on the customer