(CDPE-2170) Only create cd4pe docker network when using mysql db
Prior to this commit, we created the cd4pe docker network in all cases.
This meant that customers had to ensure they were not already using the
class B network this defaults to consuming.  This additional overhead is
only needed when a managed mysql DB is used, which is not the default
for new installations and thus unlikely.

With this commit, we now check whether or not mysql is in use during
init and only create the additional docker network if so.  The logic
used to determine that mysql was already in place was down in the
database code, so that needed to be pulled up to the init level so it
could be leveraged prior to creating the network.