(maint) Add back in docker network for tests
Previous to this commit, the docker network was removed for postgres
based DB's since it is a system package install normally. However in our
acceptance tests we use docker in a container. With the network removal,
we were unable to stand up the postgres container.
This commit injects the network back into the test manifest.