Update puppet apply check logic
Prior to this commit, if you run the cd4pe install task from the PE console more than once the second run will fail because it tries to include cdpe::repo which doesn't work via puppet apply. 

After this commit, we update the logic to hinge on whether we are passing the cd4pe_multimodule_package fact like the task does.