CDPE-3016 removes params, etc to install distelli agent
In CD4PE 3.4.0, we are beginning the use of the Puppet agent as the primary job runner. This means phasing the distelli agent out hard and as part of that process, we should no longer offer to automatically install the Distelli agent as part of CD4PE installation. This change removes the params, methods, tests, and dependences required for the module to be able to do this.