(CDPE-2800) tooling for creating manual test configs
This is a POC for creating CD4PE VMs that are preconfigured for our
manual testing efforts. First draft will configure Postgres as the DB,
with a selectable object store and possibly SSL. VCS will require more
work, as we currently don't have endpoints or tasks to support that
automation yet. It requires BOLT 2.x.