(CDPE-2800) Add scripts to create and configure manual test CD4PE VMs
This commit adds scripts that create and configure object storage and
enable SSL for freshly provisioned CD4PE VMs. It also creates the root
account and sets a trial license.

Use of this script requires installation of the 1Password `op` utility,
here: https://support.1password.com/command-line-getting-started/,
and full docs are here: https://support.1password.com/command-line/

Environmental setup

This was developed and tested using Ruby 2.4.1 (because thats what we
use in Jenkins). Use any Ruby environment tool (rbenv, rvm) to use that
or a higher version, and perform a `bundle install --path .vendor`.

Invoking

To specify the repo and version of the image to install, set the same
environment vars as the build system, CD4PE_IMAGE and CD4PE_VERSION;
the image repo defaults to `artifactory.delivery.puppetlabs.net/cd4pe-dev`

No arguments are required. By default, it will create a VM using the
`disk` object-store, and not enable SSL. To modify this behaviour, use
the following switches:

-o|--object-store disk|artifactory|s3  to specify the object-store
-s|--ssl  to configure SSL
-p|--no-po-check  to disable the 1Password `po` tool sanity check

The `po` tool uses a 30 minute token, and occasionally gets confused
when checking token status; this short-circuits the check, but a valid
token is still required for the tool itself to work. I created the
following Bash alias to make it simpler to re-up the token:

`alias optoken='eval $(op signin puppet)'`

Caveats & implementation notes

This script will create local `inventory.yaml` and `params.json` files
in the .../puppetlabs-cd4pe directory, and remove any existing ones.
The `inventory.yaml` file is left for further bolt runs; the `params.json`
file contains sensitive information and is removed upon a successful
run.

In order to get a current version of the `puppetlabs-cd4pe` module, it
pulls from the master branch in Github.