(CDPE-3131) Align pdb timeout env variable with backend code
Prior to this commit, changing the PDB timeout caused the module to add
PUPPET_DB_CONNECTION_TIMEOUT_SEC to the env file for cd4pe.
Unfortunately, this is not the name of the variable the backend code is
looking for, so the setting is never honored.

With this commit, the variable is renamed to
PUPPETDB_CONNECTION_TIMEOUT_SEC to match what the code is looking for at
https://github.com/puppetlabs/PipelinesInfra/blob/3.5.0-release/src/main/java/com/puppet/enterprise/puppetdb/PuppetDBClient.java#L46.