(CDPE-3062) Read test config from disk instead of 1Password
The user needs to get the file from 1Password manually and place it in
their $HOME directory before calling configTestVm.sh. This follows the
same pattern we have of putting the cdpe-test-credentials.json file in
your $HOME directory. Soon both of these files will be merged.

This removes a barrier to getting this stuff running in a Docker
container.