(maint) Fix potential SSH issues and add Gemfile.lock note
Previously we were mounting in and using the entire ~/.ssh/ directory,
which can lead to problems depending on what you have in there when
you run the script.

We only need the SSH key though, so instead just mount in that
individual file instead of the entire directory.

Also add a note to the README about Gemfile.lock causing problems and
how to fix it.