(CDPE-3231) Add OS family check for cd4pe installation
Prior to this commit, it was possible to attempt an installation of
cd4pe on an agent with a different OS release than the PE master.  This
installation attempt would fail, generally looking for postgresql
packages with an error that would likely not lead a user to suspect the
actual cause of the problem.

With this commit, we will now compare the OS family to make sure they
are the same before continuing.