(CDPE-3234) Remove pipelines dependency
The pipeline modules dependency creates conflicting version dependencies that cannot be resolved.  Instead of updating the dependencies of the pipelines module it seems reasonable to remove the dependency altogether as it was used for optional functionality and is also no longer needed with the addition of puppet agent based job hardware.