(maint) Use forge for fixtures
Previous to this commit, the fixtures file was pulling #master for all
the modules we use. This could lead to tests failing if #master of a
module repo is broken.
This commit changes it to use forge for all modules that are published
to the forge. There are still a handful of PE specific ones / helpers
that are internal only and not published to the forge. Those are left to
 #master for now.