(CDPE-3310) Fix discover_pe_credentials task workspace parameter
Prior to this commit, the discover_pe_credentials task definition
included a "workspace", but the task itself never used it.  Meanwhile
the discover_pe_credentials function in the CD4PE client code expected
it.

With this commit, the task is updated to use the workspace parameter and
send it through to the client code.