(maint) Pin to facter 2.x
Previous to this commit, we were pulling in the latest facter gem.
However that was leading to failures in CI as it was causing catalog
compilation failures with 4.x, and no useful debug logging at the time
to figure out why.
For now, pin to 2.x to get CI working again while we investigate the
root cause to remove this pinning.