(CDPE-3560) Gut manifests/Puppet api
In this commit, the work to gut the module begins with paring the manifests down to three: init, root_config, and impact_analysis. Users that have installed CD4PE with the module in the past can now run the module to undo most everything it did. init.pp has been changed to be the Great Undoer of things, including making sure all of the file resources that may have been present are now absent, the docker runs are absent, and based on your db provider, all database resources are absent. The root_config class's only purpose now is to make sure there are no more files in /home/distelli.