(CDPE-4193) Respect uri path when hitting cd4pe backend
Prior to this change, if the web_ui_endpoint supplied by CD4PE contained
a path after the port, we ignored it and simply attempted to hit the
server/port. This historically worked just fine, as CD4PE never
specified a path after the port; however, in CD4PE 4.5.0, we introduced
a path prefix that all traffic to CD4PE has to respect.

With this change, we add logic to respect the URI path if one exists,
allowing us to use the puppetlabs-cd4pe module with CD4PE 4.5.0+.