fix: Properly handle non-SSL endpoints
This fixes the `SSL_connect returned=1 errno=0 state=error: wrong version number` error that happens when you force Ruby to use SSL while the target endpoint is only providing HTTP and no SSL. CD4PE 3.x runs on HTTP by default and many customers have not setup for it, causing Tasks in v3.2.1 of the `cd4pe` module to fail against the HTTP endpoint.
The change ensures that SSL attributes are only set for the connection if the URI schema is `https`.