(CDPE-4881) account for /cd4pe prefix in default web_ui_endpoint behavior
This updates the default for the web_ui_endpoint to add /cd4pe onto the end of the path, as required by teams ui. This only changes the default behavior if the user does not supply the param themself, so as not to break users who have already updated it accordingly on their own.