(CDPE-4883) clarify generic login error response
Previously the generic error handler that'd catch all non-401 error states when the client was setting the auth cookie would swallow the error response and instead return an error about invalid login credentials. However this was misleading as invalid credentials would result in a 401, and the handler was catching all error responses except 401s. This adjusts it to instead return the error status and body.