(CDPE-5783) Prepare for automatic pushes from PipelinesInfra
Delete everything but the .git/ directory so the only contents of this
branch will be the CD4PE module tarball built from the PipelinesInfra
repository.

This branch is not intended for development, it is for internal users to
get our module. Development of the module happens in PipelinesInfra
deploy/ directory, and on merge to main a GitHub Action will push the
module to this repository.